+++ suggestion/__tests__/integration.test.ts
@@ -1,0 +1,1 @@
+import { suggestRecipes } from '../src/controller/suggestRecipe';
