+++ suggestion/src/routes/suggestion.ts
@@ -1,9 +1,23 @@
 import * as express from 'express';
-import { suggestRecipeController } from '../controller/suggestRecipe';
+import { suggestRecipes } from '../controller/suggestRecipe';
 
 const router = express.Router();
 
 // POST /suggestion?ingredients=xyz
-router.post('/', suggestRecipeController);
+router.post('/', (req: express.Request, res: express.Response) => {
+    const httpBody = req.body;
+    const numIngredients = httpBody.queryIngredients.length;
+
+    const testThreshold = 0.5;
+
+    const IDs = [];
+
+    for (let i = 0; i < numIngredients; i++) {
+        IDs.push(httpBody.queryIngredients[i].databaseID);
+    }
+
+    const recipeHashes = suggestRecipes(IDs, testThreshold);
+    res.status(200).send('Result: ' + recipeHashes);
+});
 
 module.exports = router;
